+++ src/data/response.kt
@@ -37,6 +37,8 @@ data class ResponseAuthUser(
     val name: String,
     val allianceName: String,
     val allianceTicker: String,
+    val csrfHeaderKey: String,
+    val csrfToken: String,
 )
 
 data class ResponseGates(