+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 ## 0.2.0
 
-yyyy-mm-dd
+2020-09-20
 
 - [#3][i3] Added support for PostgreSQL, MySQL, MariaDB, SQLite and H2 (embedded mode) databases.
 - User is logged off from the frontend if the backend responds with a "Forbidden" error.